sql: fix integer overflow in built-in functions
This patch replaces the type for some int and uint32_t values with
size_t to avoid problems with integer overflow.

Closes tarantool/security#119

NO_DOC=bugfix

+++ src/box/sql/vdbe.c
@@ -129,7 +129,7 @@ int sql_sort_count = 0;
  * help verify the correct operation of the library.
  */
 #ifdef SQL_TEST
-int sql_max_blobsize = 0;
+size_t sql_max_blobsize = 0;
 static void
 updateMaxBlobsize(Mem *p)
 {
